backend/requirements.txt
django
djangorestframework
django-ninja
django-ninja-jwt
pytest
pytest-django
ruff
python-dotenv
